hooked up first vc
