--- Fitbod-Challenge/UI/OneRepMaxViewController.swift
@@ -1,26 +1,0 @@
-//
-//  OneRepMaxViewController.swift
-//  Fitbod-Challenge
-//
-//  Created by Ambuj Punn on 7/30/18.
-//  Copyright © 2018 Ambuj Punn. All rights reserved.
-//
-
-import UIKit
-
-class OneRepMaxViewController: UIViewController {
-    let workoutData: WorkoutData? = WorkoutData.shared
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-
-    }
-
-    override func didReceiveMemoryWarning() {
-        super.didReceiveMemoryWarning()
-        // Dispose of any resources that can be recreated.
-    }
-
-
-}
-
