+++ Fitbod-Challenge/UI/Views/ExerciseTableViewCell.swift
@@ -1,0 +1,25 @@
+//
+//  ExerciseTableViewCell.swift
+//  Fitbod-Challenge
+//
+//  Created by Ambuj Punn on 8/1/18.
+//  Copyright © 2018 Ambuj Punn. All rights reserved.
+//
+
+import UIKit
+
+class ExerciseTableViewCell: UITableViewCell {
+
+    @IBOutlet weak var nameLabel: UILabel!
+    @IBOutlet weak var weightLabel: UILabel!
+    
+    override func awakeFromNib() {
+        super.awakeFromNib()
+    }
+
+    override func setSelected(_ selected: Bool, animated: Bool) {
+        super.setSelected(selected, animated: animated)
+
+    }
+
+}
